final commit of version 1.7

+++ VAMP_NBP/VAMP_Parser/src/de/cebitec/vamp/parser/Job.java
@@ -13,12 +13,10 @@ public interface Job {
 
     public File getFile();
 
-//    public ParserI getParser();
-
     public String getDescription();
 
     public Timestamp getTimestamp();
 
-    public Long getID();
+    public int getID();
 
 }
